refactor: rename to `SelectableText` instead of `Text`
This avoids API breakage
